feat(Calendar): [BLA-694] move eventDetails interfaces to the interaction interfaces
