+++ packages/core-components/src/utils/interfaces/interaction.interface.d.ts
@@ -18,6 +18,12 @@ export interface NextMonth<T = any> {}
 export interface CalendarViewChangedEventDetail<T = CalendarView> {
   value: T;
 }
+export interface MonthSelectedEventDetail<T = number> {
+  value: T;
+}
+export interface YearSelectedEventDetail<T = number> {
+  value: T;
+}
 export interface ColumnSortChangeEventDetail {
   sortedColumn: string;
   sortDirection: string;
